refactor(v1): made v1 code reuseable for v2 - all still compliant
